Fix Micro USB connector
It seems that at least 3 versions of datasheets exists. One of them
seems to have completely broken some of the dimensions. This led to
misplacement on PCB and conflict with Hirose connector. The other
datasheets differ in electrical characteristics. So in case of
mechanical dimensions, we can try believing them. Also photo shared by
LCSC confirms the theory that they are right. So, for now, we assume
that official site shares the correct datasheets. Chinese is more
trustworthy in terms of minor characteristics.