Add new Micro USB footprint
It was overlooked that element was swapped in BOM from *4BH83 to *4BH81.
As a result former was used till now. The only difference, fortunately
is throught hole mounting point, so only minor fix in routing was
needed.

Still there is weird issue with one ground pad, that does not want to
update itself.