Fix width of main power rail
By mistake it was normal trace width near Hirose connector, while it
should be 1.2 mm. To fix that with as little change as possible, set it
to 2/3 of its required width after first Hirose pin trace is branched,
then to 1/3 after second branch. This should not harm as whatever
current it needs to carry will be split between 6 traces near the
connector.

This fixes #2.