Swap sides for power and reset button
It is more convenient to have reset on top, as it is not at the edge of
board. Also it decreases a risk of damage to CM4, as it may happen
someone will try to reset with conductive, metal thing and touch
something on CM4. Having power button on bottom (CM4) side does not pose
that risk as it is exactly at the edge.