Add dummy symbol and footprint for Hirose
CM4 is two-unit symbol with one footprint, designated Module1, which is
useless for fabrication. Add two dummy symbols marking both Hirose
connectors as ordinary J? parts. This allows them to appear on
silkscreen automagically and also be in BOM as proper entries. Only
unlinked to real pads, but who cares.